__dict_delrec: Changed so that no “use after free warning” is issued by gcc
Technically, the warning was correct:
The address of the deleted element was returned after free() was called and used by the calling function to check whether the first element in struct_dict was deleted by comparing this address with the address of the first element.
Since this was only used to determine whether the first element was affected or not, this did not result in a security vulnerability.